Move CalculatedModelValue out of ModelContainer
This was sort of a separate thing, only used by DefaultConfiguration. Pull it out into its own class
