+++ platforms/core-configuration/model-core/src/main/java/org/gradle/internal/model/ModelContainer.java
@@ -30,8 +30,6 @@ public interface ModelContainer<T> {
     /**
      * Runs the given function to calculate a value from the public mutable model. Applies best effort synchronization to prevent concurrent access to a particular project from multiple threads.
      * However, it is currently easy for state to leak from one project to another so this is not a strong guarantee.
-     *
-     * <p>It is usually a better option to use {@link #newCalculatedValue(Object)} instead of this method.</p>
      */
     <S extends @Nullable Object> S fromMutableState(Function<? super T, ? extends S> factory);
 
@@ -57,8 +55,4 @@ public interface ModelContainer<T> {
      */
     boolean hasMutableState();
 
-    /**
-     * Creates a new container for a value that is calculated from the mutable state of this container, and then reused by multiple threads.
-     */
-    <S> CalculatedModelValue<S> newCalculatedValue(@Nullable S initialValue);
 }
