+++ platforms/software/dependency-management/src/main/java/org/gradle/api/internal/ConfigurationServicesBundle.java
@@ -29,6 +29,7 @@ import org.gradle.api.problems.internal.ProblemsInternal;
 import org.gradle.api.provider.ProviderFactory;
 import org.gradle.internal.model.CalculatedValueContainerFactory;
 import org.gradle.internal.operations.BuildOperationRunner;
+import org.gradle.internal.resources.ProjectLeaseRegistry;
 import org.gradle.internal.service.scopes.Scope;
 import org.gradle.internal.service.scopes.ServiceScope;
 
@@ -55,4 +56,5 @@ public interface ConfigurationServicesBundle {
     AttributeDesugaring getAttributeDesugaring();
     ResolveExceptionMapper getExceptionMapper();
     ProviderFactory getProviderFactory();
+    ProjectLeaseRegistry getProjectLeaseRegistry();
 }
