+++ subprojects/core/src/main/java/org/gradle/api/internal/initialization/StandaloneDomainObjectContext.java
@@ -21,7 +21,6 @@ import org.gradle.api.internal.project.ProjectIdentity;
 import org.gradle.api.internal.project.ProjectInternal;
 import org.gradle.api.internal.project.ProjectState;
 import org.gradle.groovy.scripts.ScriptSource;
-import org.gradle.internal.model.CalculatedModelValue;
 import org.gradle.internal.model.ModelContainer;
 import org.gradle.util.Path;
 import org.jspecify.annotations.Nullable;
@@ -203,44 +202,4 @@ public abstract class StandaloneDomainObjectContext implements DomainObjectConte
         return false;
     }
 
-    @Override
-    public <T> CalculatedModelValue<T> newCalculatedValue(@Nullable T initialValue) {
-        return new CalculatedModelValueImpl<>(initialValue);
-    }
-
-    private static class CalculatedModelValueImpl<T> implements CalculatedModelValue<T> {
-        private volatile T value;
-
-        CalculatedModelValueImpl(@Nullable T initialValue) {
-            value = initialValue;
-        }
-
-        @Override
-        public T get() throws IllegalStateException {
-            T currentValue = getOrNull();
-            if (currentValue == null) {
-                throw new IllegalStateException("No value is available.");
-            }
-            return currentValue;
-        }
-
-        @Override
-        public T getOrNull() {
-            return value;
-        }
-
-        @Override
-        public void set(T newValue) {
-            value = newValue;
-        }
-
-        @Override
-        public T update(Function<T, T> updateFunction) {
-            synchronized (this) {
-                T newValue = updateFunction.apply(value);
-                value = newValue;
-                return newValue;
-            }
-        }
-    }
 }
